RDCC-3011: Added a test case for validating number of params passed to the endpoint. Also, corrected a few existing tests.
